Upgrade Fuel ecosystem dependencies (#999)

+++ packages/fuel-indexer/src/service.rs
@@ -323,7 +323,7 @@ async fn create_service_task(
 async fn get_start_block(
     conn: &mut IndexerConnection,
     manifest: &Manifest,
-) -> Result<u64, IndexerError> {
+) -> Result<u32, IndexerError> {
     let Manifest {
         namespace,
         identifier,
